VkChatBot/requirements.txt
certifi==2020.12.5
chardet==4.0.0
coverage==5.4
idna==2.10
Pillow==8.1.0
requests==2.25.1
six==1.15.0
urllib3==1.26.3
vk-api==11.9.1
